+++ nodeserver/sharedState.js
@@ -1,4 +1,0 @@
-// sharedState.js
-module.exports = {
-    timeoutOngoing: false
-};