+++ CHANGELOG.md
@@ -5,15 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] 2023-12-08
+## [Unreleased] 2023-12-09
 ### Changed
 - Migration of mongoClient, buildUrlMap and mqttPublisher to TS
 - Migration of authMiddleware and socketConfig to TS
 - Migration of rateLimiter and getTaskEnabler to TS
 - Migration of switchTaskEnabler to TS
+- Migration of mqttClient to TS
+- Migration of authMiddlewareSocket to TS
+- Migration of sharedState to TS
 
 ### Added
 - id and name field for frontend
+- id and name field to Settingspage GPT Request
 
 ## [v15.3.0] 2023-12-08
 ### Removed